Created TickerRegistry Moved registries into new file Renamed PluginManager to PluginLoader
