+++ abacura/config.py
@@ -66,12 +66,12 @@ class Config(Plugin):
 
         log(f"@conf called with full '{full}' and reload '{reload}'")
         if full or self.app.session == "null":
-            conf = escape(self.manager.config.config.as_string())
+            conf = escape(self.config.config.as_string())
         else:
-            conf = escape(self.manager.config.config[self.app.session].as_string())
+            conf = escape(self.config.config[self.app.session].as_string())
         
         panel = Panel(conf, highlight=True)
-        tl = self.manager.tl
+        tl = self.session.tl
         tl.markup = True
         tl.write(panel)
         tl.markup = False
